pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "awsug-cert-generator"
version = "1.0.0"
description = "AWS User Group Certificate Generator"
requires-python = ">=3.9"
dependencies = [
    "reportlab>=4.2.5",
    "Pillow>=10.4.0",
    "PyYAML>=6.0.2",
    "Flask>=3.0.0",
]

[project.scripts]
cert-generator = "cli:main"

[tool.setuptools]
py-modules = ["cli", "app"]

[tool.setuptools.packages.find]
where = ["."]
